+++ configs/src/index.ts
@@ -1,3 +1,2 @@
 export * from './networks';
-export * from './types';
-export * from './deployments/index';
+export * from './deployments';
